DONE - Transfer - add to score DONE - Disrupter weapon DONE - Droid weapon firing rate and recharge DONE - Droid takeover countdown

+++ hdr/game/database.h
@@ -14,7 +14,7 @@ typedef struct
 	int         chanceToShoot;
 	int         bulletDamage;
 	bool        disrupterImmune;
-	double      rechargeTime;
+	float       rechargeTime;
 	int         tokenCount;
 	int         imageIndex;
 	std::string dbImageFileName;
